Changed long to int for port cv min and max and refactored names BugzId: 940

+++ com-mmxlabs-models-lng-transformer/com.mmxlabs.models.lng.transformer/src/com/mmxlabs/models/lng/transformer/LNGScenarioTransformer.java
@@ -436,8 +436,8 @@ public class LNGScenarioTransformer {
 
 		for (final Port ePort : portModel.getPorts()) {
 			final IPort port;
-			final long minCv = ePort.isSetMinCvValue() ? OptimiserUnitConvertor.convertToInternalConversionFactor(ePort.getMinCvValue()) : 0;
-			final long maxCv = ePort.isSetMaxCvValue() ? OptimiserUnitConvertor.convertToInternalConversionFactor(ePort.getMaxCvValue()) : Long.MAX_VALUE;
+			final int minCv = ePort.isSetMinCvValue() ? OptimiserUnitConvertor.convertToInternalConversionFactor(ePort.getMinCvValue()) : 0;
+			final int maxCv = ePort.isSetMaxCvValue() ? OptimiserUnitConvertor.convertToInternalConversionFactor(ePort.getMaxCvValue()) : Integer.MAX_VALUE;
 			if (ePort.getLocation() != null) {
 				final Location loc = ePort.getLocation();
 				port = builder.createPort(ePort.getName(), !ePort.isAllowCooldown(), cooldownCalculators.get(ePort), (float) loc.getLat(), (float) loc.getLon(), ePort.getTimeZone(), minCv, maxCv);
